requirements.txt
requests>=2.28.1,<3
python-dotenv
ocean-lib==2.2.1
# Add explicit brownie requirement needed by ocean-lib
eth-brownie>=1.19.2